memory/mem_encrypt: fix reboot loop and build modules in runtest
There's no quit condition in the runtest.sh, here addes a flag in the
flag file as quit condition. And compile the module inside the
runtest.sh instead of Makefile, because we need to install the
corresponding kernel-{rt,64k,debug-*}-devel.

+++ general/memory/function/mem_encrypt/runtest.sh
@@ -38,9 +38,29 @@ krelease=$(rpm -qf --qf "%{release}\n" /boot/vmlinuz-$(uname -r))
 
 rlJournalStart
 	rlPhaseStartTest
-	if grep mem_encrypt=on /proc/cmdline ; then
-		rlLog "already set"
-		grep "\<SME\>" <(journalctl -kb) || return 0
+	if test -f $firstboot && grep "done" $firstboot; then
+		grep "mem_encrypt=on" /proc/cmdline
+		rlRun "rm -f $firstboot"
+		exit 0
+	elif grep mem_encrypt=on /proc/cmdline ; then
+		rlLog "already set mem_encrypt=on"
+		if ! test -f $firstboot; then
+			rlLog "by default"
+			rlRun "touch $firstboot"
+			skip_cleanup_cmdline=1
+			echo "skip_cleanup_cmdline" > $firstboot
+		fi
+		grep "\<SME\>" <(journalctl -kb) || {
+			rlLog "sme not enabled in bios or not supported, skip test ..."
+			report_result "sme not enabled in bios" SKIP
+			((skip_cleanup_cmdline == 1)) || {
+				rlRun "grubby --remove-args mem_encrypt=on --update-kernel DEFAULT"
+				touch $firstboot
+				echo "done" >> $firstboot
+				rhts-reboot
+			}
+			exit 0
+		}
 	elif [ ! -e $firstboot ]; then
 		rlRun "grubby --args mem_encrypt=on --update-kernel DEFAULT"
 		rlRun "touch $firstboot"
@@ -48,12 +68,15 @@ rlJournalStart
 	fi
 	rpm -q ${kname}-devel-${kversion}-${krelease} || rlRpmInstall ${kname}-devel $kversion $krelease "$(uname -m)"
 	rpm -q ${kname}-devel-${kversion}-${krelease} || rlDie "no ${kname}-devel package available"
+	rlRun "make -C sme_module" 0
+	rlRun "dmesg -C"
 	rlRun "insmod sme_module/sme_test.ko" 1
 	rlRun "dmesg | grep SMEtest | awk '{print \$5,\$6,\$7,\$8,\$9,\$10,\$11,\$12}' | grep -v deadbeef" 0
 	rlRun "dmesg -C"
 	if [ -e $firstboot ]; then
+		grep skip_cleanup_cmdline $firstboot && echo "reserve mem_encrypt=on in cmdline" && rm -f $firstboot && exit 0
 		rlRun "grubby --remove-args mem_encrypt=on --update-kernel DEFAULT"
-		rlRun "rm $firstboot"
+		rlRun "echo done > $firstboot"
 		rhts-reboot
 	fi
 	rlPhaseEnd
